extract firestore-collection & cleanup.
