+++ src/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <title>Example Interactives</title>
-    <link rel="stylesheet" href="../index.css" />
+    <link rel="stylesheet" href="index.css" />
   </head>
   <body>
     <div id="app">
